backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.7.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-groq>=0.1.5
langchain-core>=0.2.0
httpx>=0.27.0
sse-starlette>=2.0.0
python-multipart>=0.0.9
